Imported Upstream version 1.2.4

+++ src/dnsmasq-manager/nm-dnsmasq-utils.c
@@ -18,7 +18,8 @@
  * Copyright (C) 2013 Red Hat, Inc.
  */
 
-#include <config.h>
+#include "nm-default.h"
+
 #include <string.h>
 #include <arpa/inet.h>
 
@@ -33,7 +34,7 @@ nm_dnsmasq_utils_get_range (const NMPlatformIP4Address *addr,
                             char **out_error_desc)
 {
 	guint32 host = addr->address;
-	guint32 prefix = addr->plen;
+	guint8 prefix = addr->plen;
 	guint32 netmask = nm_utils_ip4_prefix_to_netmask (prefix);
 	guint32 first, last, reserved;
